+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Fallback to JDI `toString` when there is no model class available [#1587](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1587)
 - Auto-generate `toString` for atomic unique attributes for non-preconfigured models [#1588](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1588)
 - Inform about obsolete and not available anymore Item Types [#1589](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1589)
+- Enhanced support of the `localized` fields by showing the values in all locales [#1590](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1590)
 
 ### `Type System` enhancements
 - Introduced custom icon for `dynamic` attributes [#1582](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1582)